Moving TrelloTemplate to org.trello4j.core
